+++ js/api.js
@@ -1,5 +1,4 @@
-//import { createMarkerGroup } from './form.js';
-import { showAlert } from './cards.js';
+import { showAlert } from './util.js';
 
 
 const getData = (onSuccess) => {
